feat: dedicated getGoPath for nacl.
