+++ interp/gopath_nacl.go
@@ -1,0 +1,13 @@
+// +build nacl
+
+package interp
+
+import "go/build"
+
+func getGoPath(options Options) string {
+	if options.GoPath != "" {
+		return options.GoPath
+	}
+
+	return build.Default.GOPATH
+}
